# docs/openapi.yaml
openapi: 3.1.0
info:
  title: mKey daemon API
  version: 1.0.0
  description: |
    HTTP API демона mKey (SPEC §9, модуль internal/api).

    Транспорты:
    - Unix-сокет `$XDG_RUNTIME_DIR/mkey/mkey.sock` (права 0600) — для CLI, без токена;
    - `http://127.0.0.1:17420` — для веб-интерфейса: заголовок `Host` только 127.0.0.1/localhost,
      для изменяющих запросов `Origin` — только страницы mKey, токен из `$XDG_RUNTIME_DIR/mkey/token`
      в заголовке `Authorization: Bearer …` или cookie `mkey_session` (ставится при входе по `/?t=<токен>`).

    Сообщения ошибок переводятся на язык из `Accept-Language` (ru, en).
servers:
  - url: http://127.0.0.1:17420
security:
  - bearer: []
  - cookie: []
paths:
...
  /api/v1/plugins:
    get:
      summary: "Плагины (ADR-0029): {plugins: [{id, version, kind, name, description, dir, system, permissions, active, state, error, actions, conditions, triggers, templates}], dir}"
      responses: { "200": { description: "{plugins, dir}" }, "503": { $ref: "#/components/responses/Error" } }
  /api/v1/plugins/install:
    post:
      summary: "Установить плагин (выключенным): JSON {path} — папка или .zip на этом компьютере, либо тело application/zip — загруженный архив"
      responses:
        "200": { description: Сведения о плагине }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/plugins/{id}/{action}:
    post:
      summary: "enable | disable — включить или выключить; список включённых сохраняется в config.yaml (modules.plugins.active)"
      responses: { "200": { description: "{ok}" }, "404": { $ref: "#/components/responses/Error" } }
  /api/v1/plugins/{id}:
    delete:
      summary: "Удалить плагин пользователя (системный — 400 api.plugin_system)"
      responses: { "200": { description: "{ok}" }, "404": { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
    cookie: { type: apiKey, in: cookie, name: mkey_session }
  responses:
    Error:
      description: Ошибка в едином формате
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string, example: dsl.unknown_key_hint }
                  message: { type: string, example: "неизвестная клавиша «Mous0» — возможно, вы имели в виду «Mouse0»?" }
                  details:
                    type: object
                    properties:
                      pos:
                        type: object
                        properties:
                          line: { type: integer }
                          col: { type: integer }
                      args: { type: object, additionalProperties: { type: string } }
...